% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodmap-package.R
\docType{package}
\name{foodmap-package}
\alias{foodmap}
\alias{foodmap-package}
\title{foodmap: Map Australia's Best Restaurants from Multiple Guides}
\description{
Scrapes restaurant guides (Broadsheet, Gourmet Traveller, Time Out, AGFG, Urban List, Good Food Guide), geocodes venues via the Google Places API, and exports KML files for import into Google My Maps.
}
\author{
\strong{Maintainer}: Peter Owen \email{peter@bzi.com.au}

}
\keyword{internal}
